// src/types/api.types.ts
export interface MarketIndicator {
  id: number;
  name: string;
  title: string;
  p: string;
  h: string;
  l: string;
  o: string;
  d: string;
  dp: number;
  dt: 'high' | 'low';
  t: string;
  updated_at: string;
}